[Fix] Improve consensus validation before execution.
