+++ src/AElf.Kernel.Consensus.AEDPoS/AssemblyInfo.cs
@@ -1,6 +1,4 @@
 using System.Runtime.CompilerServices;
 
-#if DEBUG
 [assembly: InternalsVisibleTo("AElf.Kernel.Tests")]
 [assembly: InternalsVisibleTo("AElf.Contracts.Consensus.AEDPoS.Tests")]
-#endif
